refactor: move `models` to top-level
